Crate google_appengine1_beta4

This documentation was generated from appengine crate version 1.0.0+20160802, where 20160802 is the exact revision of the appengine:v1beta4 schema built by the mako code generator v1.0.0.

Everything else about the appengine v1_beta4 API can be found at the official documentation site. The original source code is on github.

Features

Handle the following Resources with ease from the central hub ...

apps

create, get, locations get, locations list, modules delete, modules get, modules list, modules patch, modules versions create, modules versions delete, modules versions get, modules versions instances debug, modules versions instances delete, modules versions instances get, modules versions instances list, modules versions list, modules versions patch, operations get and operations list

Not what you are looking for ? Find all other Google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

creates Method Builders which in turn allow access to individual Call Builders

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

All structures are marked with applicable traits to further categorize them and ease browsing.

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.apps().modules_patch(...).doit()
let r = hub.apps().create(...).doit()
let r = hub.apps().modules_versions_instances_delete(...).doit()
let r = hub.apps().modules_versions_instances_debug(...).doit()
let r = hub.apps().modules_delete(...).doit()
let r = hub.apps().operations_get(...).doit()
let r = hub.apps().modules_versions_patch(...).doit()
let r = hub.apps().modules_versions_delete(...).doit()
let r = hub.apps().modules_versions_create(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
google-appengine1_beta4 = "*"

A complete example

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_appengine1_beta4 as appengine1_beta4;
use appengine1_beta4::Module;
use appengine1_beta4::{Result, Error};
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use appengine1_beta4::Appengine;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Appengine::new(hyper::Client::new(), auth);
// As the method needs a request, you would usually fill it with the desired information
// into the respective structure. Some of the parts shown here might not be applicable !
// Values shown here are possibly random and not representative !
let mut req = Module::default();
 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.apps().modules_patch(req, "appsId", "modulesId")
             .migrate_traffic(true)
             .mask("sit")
             .doit();
 
match result {
    Err(e) => match e {
        // The Error enum provides details about what exactly happened.
        // You can also just use its `Debug`, `Display` or `Error` traits
         Error::HttpError(_)
        |Error::MissingAPIKey
        |Error::MissingToken(_)
        |Error::Cancelled
        |Error::UploadSizeLimitExceeded(_, _)
        |Error::Failure(_)
        |Error::BadRequest(_)
        |Error::FieldClash(_)
        |Error::JsonDecodeError(_, _) => println!("{}", e),
    },
    Ok(res) => println!("Success: {:?}", res),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

Uploads and Downloads

If a method supports downloads, the response body, which is part of the Result, should be read by you to obtain the media. If such a method also supports a Response Result, it will return that by default. You can see it as meta-data for the actual media. To trigger a media download, you will have to set up the builder by making this call: .param("alt", "media").

Methods supporting uploads can do so using up to 2 different protocols: simple and resumable. The distinctiveness of each is represented by customized doit(...) methods, which are then named upload(...) and upload_resumable(...) respectively.

Customization and Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

Optional Parts in Server-Requests

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifiable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are moved

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Structs

ApiConfigHandler
Google Cloud Endpoints configuration for API handlers.

ApiEndpointHandler
Uses Google Cloud Endpoints to handle requests.

AppCreateCall
Creates an App Engine application for a Google Cloud Platform project. This requires a project that excludes an App Engine application. For details about creating a project without an application, see the Google Cloud Resource Manager create project topic.

AppGetCall
Gets information about an application.

AppLocationGetCall
Get information about a location.

AppLocationListCall
Lists information about the supported locations for this service.

AppMethods
A builder providing access to all methods supported on app resources. It is not used directly, but through the Appengine hub.

AppModuleDeleteCall
Deletes the specified module and all enclosed versions.

AppModuleGetCall
Gets the current configuration of the specified module.

AppModuleListCall
Lists all the modules in the application.

AppModulePatchCall
Updates the configuration of the specified module.

AppModuleVersionCreateCall
Deploys code and resource files to a new version.

AppModuleVersionDeleteCall
Deletes an existing version.

AppModuleVersionGetCall
Gets the specified Version resource. By default, only a BASIC_VIEW will be returned. Specify the FULL_VIEW parameter to get the full resource.

AppModuleVersionInstanceDebugCall
Enables debugging on a VM instance. This allows you to use the SSH command to connect to the virtual machine where the instance lives. While in "debug mode", the instance continues to serve live traffic. You should delete the instance when you are done debugging and then allow the system to take over and determine if another instance should be started. Only applicable for instances in App Engine flexible environment.

AppModuleVersionInstanceDeleteCall
Stops a running instance.

AppModuleVersionInstanceGetCall
Gets instance information.

AppModuleVersionInstanceListCall
Lists the instances of a version.

AppModuleVersionListCall
Lists the versions of a module.

AppModuleVersionPatchCall
Updates the specified Version resource. You can specify the following fields depending on the App Engine environment and type of scaling that the version resource uses: * serving_status: For Version resources that use basic scaling, manual scaling, or run in the App Engine flexible environment. * instance_class: For Version resources that run in the App Engine standard environment. * automatic_scaling.min_idle_instances: For Version resources that use automatic scaling and run in the App Engine standard environment. * automatic_scaling.max_idle_instances: For Version resources that use automatic scaling and run in the App Engine standard environment.

AppOperationGetCall
Gets the latest state of a long-running operation. Clients can use this method to poll the operation result at intervals as recommended by the API service.

AppOperationListCall
Lists operations that match the specified filter in the request. If the server doesn't support this method, it returns UNIMPLEMENTED. NOTE: the name binding below allows API services to override the binding to use different resource name schemes, such as users/*/operations.

Appengine
Central instance to access all Appengine related resource activities

Application
An Application resource contains the top-level configuration of an App Engine application.

AutomaticScaling
Automatic scaling is based on request rate, response latencies, and other application metrics.

BasicScaling
A module with basic scaling will create an instance when the application receives a request. The instance will be turned down when the app becomes idle. Basic scaling is ideal for work that is intermittent or driven by user activity.

ContainerInfo
Docker image that is used to start a VM container for the version you deploy.

CpuUtilization
Target scaling by CPU usage.

DebugInstanceRequest
Request message for Instances.DebugInstance.

DefaultDelegate
A delegate with a conservative default implementation, which is used if no other delegate is set.

Deployment
Code and application artifacts used to deploy a version to App Engine.

DiskUtilization
Target scaling by disk usage. Only applicable for VM runtimes.

ErrorHandler
Custom static error page to be served when an error occurs.

ErrorResponse
A utility to represent detailed errors we might see in case there are BadRequests. The latter happen if the sent parameters or request structures are unsound

FileInfo
Single source file that is part of the version to be deployed. Each source file that is deployed must be specified separately.

HealthCheck
Health checking configuration for VM instances. Unhealthy instances are killed and replaced with new instances. Only applicable for instances in App Engine flexible environment.

Instance
An Instance resource is the computing unit that App Engine uses to automatically scale an application.

Library
Third-party Python runtime library that is required by the application.

ListInstancesResponse
Response message for Instances.ListInstances.

ListLocationsResponse
The response message for LocationService.ListLocations.

ListModulesResponse
Response message for Modules.ListModules.

ListOperationsResponse
The response message for Operations.ListOperations.

ListVersionsResponse
Response message for Versions.ListVersions.

Location
A resource that represents Google Cloud Platform location.

ManualScaling
A module with manual scaling runs continuously, allowing you to perform complex initialization and rely on the state of its memory over time.

MethodInfo
Contains information about an API request.

Module
A Module resource is a logical component of an application that can share state and communicate in a secure fashion with other modules. For example, an application that handles customer requests might include separate modules to handle tasks such as backend data analysis or API requests from mobile devices. Each module has a collection of versions that define a specific set of code used to implement the functionality of that module.

MultiPartReader
Provides a Read interface that converts multiple parts into the protocol identified by RFC2387. Note: This implementation is just as rich as it needs to be to perform uploads to google APIs, and might not be a fully-featured implementation.

Network
Extra network settings. Only applicable for VM runtimes.

NetworkUtilization
Target scaling by network usage. Only applicable for VM runtimes.

Operation
This resource represents a long-running operation that is the result of a network API call.

RequestUtilization
Target scaling by request utilization. Only applicable for VM runtimes.

Resources
Machine resources for a version.

ScriptHandler
Executes a script to handle the request that matches the URL pattern.

SourceReference
Reference to a particular snapshot of the source tree used to build and deploy the application.

StaticDirectoryHandler
Files served directly to the user for a given URL, such as images, CSS stylesheets, or JavaScript source files. Static directory handlers make it easy to serve the entire contents of a directory as static files.

StaticFilesHandler
Files served directly to the user for a given URL, such as images, CSS stylesheets, or JavaScript source files. Static file handlers describe which files in the application directory are static files, and which URLs serve them.

Status
The Status type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by gRPC. The error model is designed to be: - Simple to use and understand for most users - Flexible enough to meet unexpected needs # Overview The Status message contains three pieces of data: error code, error message, and error details. The error code should be an enum value of google.rpc.Code, but it may accept additional error codes if needed. The error message should be a developer-facing English message that helps developers understand and resolve the error. If a localized user-facing error message is needed, put the localized message in the error details or localize it in the client. The optional error details may contain arbitrary information about the error. There is a predefined set of error detail types in the package google.rpc which can be used for common error conditions. # Language mapping The Status message is the logical representation of the error model, but it is not necessarily the actual wire format. When the Status message is exposed in different client libraries and different wire protocols, it can be mapped differently. For example, it will likely be mapped to some exceptions in Java, but more likely mapped to some error codes in C. # Other uses The error model and the Status message can be used in a variety of environments, either with or without APIs, to provide a consistent developer experience across different environments. Example uses of this error model include: - Partial errors. If a service needs to return partial errors to the client, it may embed the Status in the normal response to indicate the partial errors. - Workflow errors. A typical workflow has multiple steps. Each step may have a Status message for error reporting purpose. - Batch operations. If a client uses batch request and batch response, the Status message should be used directly inside batch response, one for each error sub-response. - Asynchronous operations. If an API call embeds asynchronous operation results in its response, the status of those operations should be represented directly using the Status message. - Logging. If some API errors are stored in logs, the message Status could be used directly after any stripping needed for security/privacy reasons.

TrafficSplit
Traffic routing configuration for versions within a single module. Traffic splits define how traffic directed to the module is assigned to versions.

UrlDispatchRule
Rules to match an HTTP request and dispatch that request to a module.

UrlMap
URL pattern and description of how the URL should be handled. App Engine can handle URLs by executing application code, or by serving static files uploaded with the version, such as images, CSS, or JavaScript.

Version
A Version resource is a specific set of source code and configuration files that are deployed into a module.

Enums

Error

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.

Traits

CallBuilder
Identifies types which represent builders for a particular resource method

Delegate
A trait specifying functionality to help controlling any request performed by the API. The trait has a conservative default implementation.

Hub
Identifies the Hub. There is only one per library, this trait is supposed to make intended use more explicit. The hub allows to access all resource methods more easily.

MethodsBuilder
Identifies types for building methods of a particular resource type

NestedType
Identifies types which are only used by other types internally. They have no special meaning, this trait just marks them for completeness.

Part
Identifies types which are only used as part of other types, which usually are carrying the Resource trait.

ReadSeek
A utility to specify reader types which provide seeking capabilities too

RequestValue
Identifies types which are used in API requests.

Resource
Identifies types which can be inserted and deleted. Types with this trait are most commonly used by clients of this API.

ResponseResult
Identifies types which are used in API responses.

ToParts
A trait for all types that can convert themselves into a parts string

Functions

remove_json_null_values

Type Definitions

Result
A universal result type used as return for all calls.