Struct google_appengine1_beta4::api::Version

pub struct Version {
    pub api_config: Option<ApiConfigHandler>,
    pub automatic_scaling: Option<AutomaticScaling>,
    pub basic_scaling: Option<BasicScaling>,
    pub beta_settings: Option<HashMap<String, String>>,
    pub creation_time: Option<String>,
    pub default_expiration: Option<String>,
    pub deployer: Option<String>,
    pub deployment: Option<Deployment>,
    pub endpoints_api_service: Option<EndpointsApiService>,
    pub env: Option<String>,
    pub env_variables: Option<HashMap<String, String>>,
    pub error_handlers: Option<Vec<ErrorHandler>>,
    pub handlers: Option<Vec<UrlMap>>,
    pub health_check: Option<HealthCheck>,
    pub id: Option<String>,
    pub inbound_services: Option<Vec<String>>,
    pub instance_class: Option<String>,
    pub libraries: Option<Vec<Library>>,
    pub manual_scaling: Option<ManualScaling>,
    pub name: Option<String>,
    pub network: Option<Network>,
    pub nobuild_files_regex: Option<String>,
    pub resources: Option<Resources>,
    pub runtime: Option<String>,
    pub runtime_api_version: Option<String>,
    pub serving_status: Option<String>,
    pub threadsafe: Option<bool>,
    pub vm: Option<bool>,
}

A Version resource is a specific set of source code and configuration files that are deployed into a module.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

modules versions create apps (request)

modules versions get apps (response)

modules versions patch apps (request)

Fields
api_config: Option<ApiConfigHandler>

Serving configuration for Google Cloud Endpoints (https://cloud.google.com/appengine/docs/python/endpoints/).Only returned in GET requests if view=FULL is set.

automatic_scaling: Option<AutomaticScaling>

Automatic scaling is based on request rate, response latencies, and other application metrics.

basic_scaling: Option<BasicScaling>

A module with basic scaling will create an instance when the application receives a request. The instance will be turned down when the app becomes idle. Basic scaling is ideal for work that is intermittent or driven by user activity.

beta_settings: Option<HashMap<String, String>>

Metadata settings that are supplied to this version to enable beta runtime features.

creation_time: Option<String>

Time that this version was created.@OutputOnly

default_expiration: Option<String>

Duration that static files should be cached by web proxies and browsers. Only applicable if the corresponding StaticFilesHandler (https://cloud.google.com/appengine/docs/admin-api/reference/rest/v1/apps.services.versions#staticfileshandler) does not specify its own expiration time.Only returned in GET requests if view=FULL is set.

deployer: Option<String>

Email address of the user who created this version.@OutputOnly

deployment: Option<Deployment>

Code and application artifacts that make up this version.Only returned in GET requests if view=FULL is set.

endpoints_api_service: Option<EndpointsApiService>

Cloud Endpoints configuration.If endpoints_api_service is set, the Cloud Endpoints Extensible Service Proxy will be provided to serve the API implemented by the app.

env: Option<String>

App Engine execution environment to use for this version.Defaults to 1.

env_variables: Option<HashMap<String, String>>

Environment variables made available to the application.Only returned in GET requests if view=FULL is set.

error_handlers: Option<Vec<ErrorHandler>>

Custom static error pages. Limited to 10KB per page.Only returned in GET requests if view=FULL is set.

handlers: Option<Vec<UrlMap>>

An ordered list of URL-matching patterns that should be applied to incoming requests. The first matching URL handles the request and other request handlers are not attempted.Only returned in GET requests if view=FULL is set.

health_check: Option<HealthCheck>

Configures health checking for VM instances. Unhealthy instances are stopped and replaced with new instances. Only applicable for VM runtimes.Only returned in GET requests if view=FULL is set.

id: Option<String>

Relative name of the version within the module. Example: v1. Version names can contain only lowercase letters, numbers, or hyphens. Reserved names: "default", "latest", and any name with the prefix "ah-".

inbound_services: Option<Vec<String>>

Before an application can receive email or XMPP messages, the application must be configured to enable the service.

instance_class: Option<String>

Instance class that is used to run this version. Valid values are: AutomaticScaling: F1, F2, F4, F4_1G ManualScaling or BasicScaling: B1, B2, B4, B8, B4_1GDefaults to F1 for AutomaticScaling and B1 for ManualScaling or BasicScaling.

libraries: Option<Vec<Library>>

Configuration for third-party Python runtime libraries required by the application.Only returned in GET requests if view=FULL is set.

manual_scaling: Option<ManualScaling>

A module with manual scaling runs continuously, allowing you to perform complex initialization and rely on the state of its memory over time.

name: Option<String>

Full path to the Version resource in the API. Example: apps/myapp/modules/default/versions/v1.@OutputOnly

network: Option<Network>

Extra network settings. Only applicable for VM runtimes.

nobuild_files_regex: Option<String>

Files that match this pattern will not be built into this version. Only applicable for Go runtimes.Only returned in GET requests if view=FULL is set.

resources: Option<Resources>

Machine resources for this version. Only applicable for VM runtimes.

runtime: Option<String>

Desired runtime. Example: python27.

runtime_api_version: Option<String>

The version of the API in the given runtime environment. Please see the app.yaml reference for valid values at https://cloud.google.com/appengine/docs/standard//config/appref

serving_status: Option<String>

Current serving status of this version. Only the versions with a SERVING status create instances and can be billed.SERVING_STATUS_UNSPECIFIED is an invalid value. Defaults to SERVING.

threadsafe: Option<bool>

Whether multiple requests can be dispatched to this version at once.

vm: Option<bool>

Whether to deploy this version in a container on a virtual machine.