Struct google_appengine1_beta4::AppMethods

pub struct AppMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on app resources. It is not used directly, but through the Appengine hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_appengine1_beta4 as appengine1_beta4;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use appengine1_beta4::Appengine;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Appengine::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `get(...)`, `modules_delete(...)`, `modules_get(...)`, `modules_list(...)`, `modules_patch(...)`, `modules_versions_create(...)`, `modules_versions_delete(...)`, `modules_versions_get(...)`, `modules_versions_list(...)`, `modules_versions_patch(...)`, `operations_get(...)` and `operations_list(...)`
// to build up your call.
let rb = hub.apps();

Methods

impl<'a, C, A> AppMethods<'a, C, A>

fn modules_patch(&self, request: Module, apps_id: &str, modules_id: &str) -> AppModulePatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates the configuration of the specified module.

Arguments

request - No description provided.

appsId - Part of name. Name of the resource to update. For example: "apps/myapp/modules/default".

modulesId - Part of name. See documentation of appsId.

fn modules_list(&self, apps_id: &str) -> AppModuleListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists all the modules in the application.

Arguments

appsId - Part of name. Name of the resource requested. For example: "apps/myapp".

fn modules_get(&self, apps_id: &str, modules_id: &str) -> AppModuleGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the current configuration of the module.

Arguments

appsId - Part of name. Name of the resource requested. For example: "apps/myapp/modules/default".

modulesId - Part of name. See documentation of appsId.

fn get(&self, apps_id: &str) -> AppGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets information about an application.

Arguments

appsId - Part of name. Name of the application to get. For example: "apps/myapp".

fn modules_delete(&self, apps_id: &str, modules_id: &str) -> AppModuleDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes a module and all enclosed versions.

Arguments

appsId - Part of name. Name of the resource requested. For example: "apps/myapp/modules/default".

modulesId - Part of name. See documentation of appsId.

fn operations_get(&self, apps_id: &str, operations_id: &str) -> AppOperationGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the latest state of a long-running operation. Clients can use this method to poll the operation result at intervals as recommended by the API service.

Arguments

appsId - Part of name. The name of the operation resource.

operationsId - Part of name. See documentation of appsId.

fn modules_versions_patch(&self, request: Version, apps_id: &str, modules_id: &str, versions_id: &str) -> AppModuleVersionPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates an existing version. Note: UNIMPLEMENTED.

Arguments

request - No description provided.

appsId - Part of name. Name of the resource to update. For example: "apps/myapp/modules/default/versions/1".

modulesId - Part of name. See documentation of appsId.

versionsId - Part of name. See documentation of appsId.

fn operations_list(&self, apps_id: &str) -> AppOperationListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists operations that match the specified filter in the request. If the server doesn't support this method, it returns UNIMPLEMENTED. NOTE: the name binding below allows API services to override the binding to use different resource name schemes, such as users/*/operations.

Arguments

appsId - Part of name. The name of the operation collection.

fn modules_versions_delete(&self, apps_id: &str, modules_id: &str, versions_id: &str) -> AppModuleVersionDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes an existing version.

Arguments

appsId - Part of name. Name of the resource requested. For example: "apps/myapp/modules/default/versions/v1".

modulesId - Part of name. See documentation of appsId.

versionsId - Part of name. See documentation of appsId.

fn modules_versions_create(&self, request: Version, apps_id: &str, modules_id: &str) -> AppModuleVersionCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Deploys new code and resource files to a version.

Arguments

request - No description provided.

appsId - Part of name. Name of the resource to update. For example: "apps/myapp/modules/default".

modulesId - Part of name. See documentation of appsId.

fn modules_versions_get(&self, apps_id: &str, modules_id: &str, versions_id: &str) -> AppModuleVersionGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets application deployment information.

Arguments

appsId - Part of name. Name of the resource requested. For example: "apps/myapp/modules/default/versions/v1".

modulesId - Part of name. See documentation of appsId.

versionsId - Part of name. See documentation of appsId.

fn modules_versions_list(&self, apps_id: &str, modules_id: &str) -> AppModuleVersionListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists the versions of a module.

Arguments

appsId - Part of name. Name of the resource requested. For example: "apps/myapp/modules/default".

modulesId - Part of name. See documentation of appsId.