Struct google_appengine1_beta4::AutomaticScaling

pub struct AutomaticScaling {
    pub cpu_utilization: Option<CpuUtilization>,
    pub network_utilization: Option<NetworkUtilization>,
    pub min_idle_instances: Option<i32>,
    pub max_pending_latency: Option<String>,
    pub max_idle_instances: Option<i32>,
    pub disk_utilization: Option<DiskUtilization>,
    pub request_utilization: Option<RequestUtilization>,
    pub cool_down_period: Option<String>,
    pub max_total_instances: Option<i32>,
    pub max_concurrent_requests: Option<i32>,
    pub min_total_instances: Option<i32>,
    pub min_pending_latency: Option<String>,
}

Automatic scaling is the scaling policy that App Engine has used since its inception. It is based on request rate, response latencies, and other application metrics.

This type is not used in any activity, and only used as part of another schema.

Fields

cpu_utilization

Target scaling by CPU usage.

network_utilization

Target scaling by network usage.

min_idle_instances

The minimum number of idle instances that App Engine should maintain for this version. Only applies to the default version of a module, since other versions are not expected to receive significant traffic.

max_pending_latency

The maximum amount of time that App Engine should allow a request to wait in the pending queue before starting a new instance to handle it.

max_idle_instances

The maximum number of idle instances that App Engine should maintain for this version.

disk_utilization

Target scaling by disk usage.

request_utilization

Target scaling by request utilization.

cool_down_period

The amount of time that the Autoscaler should wait between changes to the number of virtual machines. Applies only to the VM runtime.

max_total_instances

Max number of instances that App Engine should start to handle requests.

max_concurrent_requests

The number of concurrent requests an automatic scaling instance can accept before the scheduler spawns a new instance. Default value is chosen based on the runtime.

min_total_instances

Minimum number of instances that App Engine should maintain.

min_pending_latency

The minimum amount of time that App Engine should allow a request to wait in the pending queue before starting a new instance to handle it.