Struct google_appengine1_beta4::UrlMap

pub struct UrlMap {
    pub static_directory: Option<StaticDirectoryHandler>,
    pub security_level: Option<String>,
    pub auth_fail_action: Option<String>,
    pub url_regex: Option<String>,
    pub script: Option<ScriptHandler>,
    pub static_files: Option<StaticFilesHandler>,
    pub api_endpoint: Option<ApiEndpointHandler>,
    pub login: Option<String>,
    pub redirect_http_response_code: Option<String>,
}

A URL pattern and description of how it should be handled. App Engine can handle URLs by executing application code, or by serving static files uploaded with the code, such as images, CSS or JavaScript.

This type is not used in any activity, and only used as part of another schema.

Fields

static_directory

Serves the entire contents of a directory as static files. This attribute is deprecated. You can mimic the behavior of static directories using static files.

security_level

Configures whether security (HTTPS) should be enforced for this URL.

auth_fail_action

For users not logged in, how to handle access to resources with required login. Defaults to "redirect".

url_regex

A URL prefix. This value uses regular expression syntax (and so regexp special characters must be escaped), but it should not contain groupings. All URLs that begin with this prefix are handled by this handler, using the portion of the URL after the prefix as part of the file path. This is always required.

script

Executes a script to handle the request that matches the URL pattern.

static_files

Returns the contents of a file, such as an image, as the response.

api_endpoint

Use API Endpoints to handle requests.

login

What level of login is required to access this resource.

redirect_http_response_code

30x code to use when performing redirects for the secure field. A 302 is used by default.