Struct google_appengine1_beta4::Version

pub struct Version {
    pub basic_scaling: Option<BasicScaling>,
    pub beta_settings: Option<HashMap<String, String>>,
    pub manual_scaling: Option<ManualScaling>,
    pub creation_time: Option<String>,
    pub vm: Option<bool>,
    pub inbound_services: Option<Vec<String>>,
    pub instance_class: Option<String>,
    pub deployment: Option<Deployment>,
    pub error_handlers: Option<Vec<ErrorHandler>>,
    pub serving_status: Option<String>,
    pub id: Option<String>,
    pub threadsafe: Option<bool>,
    pub api_config: Option<ApiConfigHandler>,
    pub nobuild_files_regex: Option<String>,
    pub network: Option<Network>,
    pub handlers: Option<Vec<UrlMap>>,
    pub health_check: Option<HealthCheck>,
    pub default_expiration: Option<String>,
    pub name: Option<String>,
    pub libraries: Option<Vec<Library>>,
    pub deployer: Option<String>,
    pub env: Option<String>,
    pub automatic_scaling: Option<AutomaticScaling>,
    pub runtime: Option<String>,
    pub env_variables: Option<HashMap<String, String>>,
    pub resources: Option<Resources>,
}

A Version resource is a specific set of source code and configuration files that are deployed into a module.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

modules versions create apps (request)

modules versions patch apps (request)

modules versions get apps (response)

Fields
basic_scaling: Option<BasicScaling>

A module with basic scaling will create an instance when the application receives a request. The instance will be turned down when the app becomes idle. Basic scaling is ideal for work that is intermittent or driven by user activity.

beta_settings: Option<HashMap<String, String>>

Metadata settings that are supplied to this version to enable beta runtime features.

manual_scaling: Option<ManualScaling>

A module with manual scaling runs continuously, allowing you to perform complex initialization and rely on the state of its memory over time.

creation_time: Option<String>

Time that this version was created. @OutputOnly

vm: Option<bool>

Whether to deploy this version in a container on a virtual machine.

inbound_services: Option<Vec<String>>

Before an application can receive email or XMPP messages, the application must be configured to enable the service.

instance_class: Option<String>

Instance class that is used to run this version. Valid values are: * AutomaticScaling: F1, F2, F4, F4_1G * ManualScaling or BasicScaling: B1, B2, B4, B8, B4_1G Defaults to F1 for AutomaticScaling and B1 for ManualScaling or BasicScaling.

deployment: Option<Deployment>

Code and application artifacts that make up this version. Only returned in GET requests if view=FULL is set.

error_handlers: Option<Vec<ErrorHandler>>

Custom static error pages. Limited to 10KB per page. Only returned in GET requests if view=FULL is set.

serving_status: Option<String>

Current serving status of this version. Only the versions with a SERVING status create instances and can be billed. SERVING_STATUS_UNSPECIFIED is an invalid value. Defaults to SERVING.

id: Option<String>

Relative name of the version within the module. Example: v1. Version names can contain only lowercase letters, numbers, or hyphens. Reserved names: "default", "latest", and any name with the prefix "ah-".

threadsafe: Option<bool>

Whether multiple requests can be dispatched to this version at once.

api_config: Option<ApiConfigHandler>

Serving configuration for Google Cloud Endpoints. Only returned in GET requests if view=FULL is set.

nobuild_files_regex: Option<String>

Files that match this pattern will not be built into this version. Only applicable for Go runtimes. Only returned in GET requests if view=FULL is set.

network: Option<Network>

Extra network settings. Only applicable for VM runtimes.

handlers: Option<Vec<UrlMap>>

An ordered list of URL-matching patterns that should be applied to incoming requests. The first matching URL handles the request and other request handlers are not attempted. Only returned in GET requests if view=FULL is set.

health_check: Option<HealthCheck>

Configures health checking for VM instances. Unhealthy instances are stopped and replaced with new instances. Only applicable for VM runtimes. Only returned in GET requests if view=FULL is set.

default_expiration: Option<String>

Duration that static files should be cached by web proxies and browsers. Only applicable if the corresponding StaticFilesHandler does not specify its own expiration time. Only returned in GET requests if view=FULL is set.

name: Option<String>

Full path to the Version resource in the API. Example: apps/myapp/modules/default/versions/v1. @OutputOnly

libraries: Option<Vec<Library>>

Configuration for third-party Python runtime libraries required by the application. Only returned in GET requests if view=FULL is set.

deployer: Option<String>

Email address of the user who created this version. @OutputOnly

env: Option<String>

App Engine execution environment to use for this version. Defaults to 1.

automatic_scaling: Option<AutomaticScaling>

Automatic scaling is based on request rate, response latencies, and other application metrics.

runtime: Option<String>

Desired runtime. Example: python27.

env_variables: Option<HashMap<String, String>>

Environment variables made available to the application. Only returned in GET requests if view=FULL is set.

resources: Option<Resources>

Machine resources for this version. Only applicable for VM runtimes.